Revert "Task/add category repository"
