+++ src/Services/ReportValidatorInterface.kt
@@ -1,4 +1,4 @@
-package src.Services.Report
+package Services
 
 interface ReportValidatorInterface {
     fun isValidYear(year: Int):Boolean
