+++ src/Services/TransactionValidatorImplementation.kt
@@ -1,4 +1,4 @@
-package src.Services.Transaction
+package Services
 
 import Utils.DateParserInterface
 import java.time.LocalDate
